Now saves newly created accounts to a text file
This newest updates now saves newly created customer accounts to a text file instead of what it used to do which was delete the account once the program closes. This will allow for other functions such as edit_account to actually do something as they have a actual account to edit.